test(duckdb): enable previously incorrectly skipped geospatial tests
