+++ ibis/backends/duckdb/tests/snapshots/test_geospatial/test_literal_geospatial_explicit/expr0/out.sql
@@ -1,6 +1,6 @@
 SELECT
-  ST_ASWKB(p) AS p
+  ST_ASWKB("p") AS p
 FROM (
   SELECT
-    ST_GEOMFROMTEXT('POINT (1 0)') AS p
+    ST_GEOMFROMTEXT('POINT (1 0)') AS "p"
 )
