+++ ibis/backends/duckdb/tests/snapshots/test_geospatial/test_literal_geospatial_inferred/shp3-0_0_1_1_2_2/out.sql
@@ -1,6 +1,6 @@
 SELECT
-  ST_ASWKB(result) AS result
+  ST_ASWKB("result") AS result
 FROM (
   SELECT
-    ST_GEOMFROMTEXT('LINESTRING (0 0, 1 1, 2 2)') AS result
+    ST_GEOMFROMTEXT('LINESTRING (0 0, 1 1, 2 2)') AS "result"
 )
