+++ ibis/backends/duckdb/tests/snapshots/test_geospatial/test_literal_geospatial_inferred/shp4-2_2_1_1_0_0/out.sql
@@ -1,6 +1,6 @@
 SELECT
-  ST_ASWKB(result) AS result
+  ST_ASWKB("result") AS result
 FROM (
   SELECT
-    ST_GEOMFROMTEXT('LINESTRING (2 2, 1 1, 0 0)') AS result
+    ST_GEOMFROMTEXT('LINESTRING (2 2, 1 1, 0 0)') AS "result"
 )
